# packages/purebackend/docs/swagger.yaml
definitions:
...
  models.R2SecretRequest:
    properties:
      access_key_id:
        type: string
      access_key_secret:
        type: string
      account_id:
        type: string
      bucket_name:
        type: string
      public_url:
        type: string
      secret_name:
        type: string
    type: object
...
  models.S3SecretRequest:
    properties:
      access_key_id:
        type: string
      access_key_secret:
        type: string
      bucket_location:
        type: string
      bucket_name:
        type: string
      secret_name:
        type: string
    type: object
...
info:
  contact:
    email: contact@pureml.com
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /org/{orgId}/secret:
    get:
      consumes:
      - '*/*'
      description: Get all secretnames of organization
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all secretnames of organization
      tags:
      - Secret
  /org/{orgId}/secret/{secretName}:
    get:
      consumes:
      - '*/*'
      description: Get secrets for secret name
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Secret Name
        in: path
        name: secretName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get secrets for secret name
      tags:
      - Secret
  /org/{orgId}/secret/{secretName}/delete:
    delete:
      consumes:
      - '*/*'
      description: Delete secrets for secret name
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Secret Name
        in: path
        name: secretName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete secrets for secret name
      tags:
      - Secret
  /org/{orgId}/secret/r2/{secretName}/test:
    get:
      consumes:
      - '*/*'
      description: Test secrets for source type r2
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Secret Name
        in: path
        name: secretName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Test secrets for source type r2
      tags:
      - Secret
  /org/{orgId}/secret/r2/connect:
    post:
      consumes:
      - '*/*'
      description: Add secrets for source type r2
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Secret
        in: body
        name: secret
        required: true
        schema:
          $ref: '#/definitions/models.R2SecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add secrets for source type r2
      tags:
      - Secret
  /org/{orgId}/secret/s3/{secretName}/test:
    get:
      consumes:
      - '*/*'
      description: Test secrets for source type s3
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Secret Name
        in: path
        name: secretName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Test secrets for source type s3
      tags:
      - Secret
  /org/{orgId}/secret/s3/connect:
    post:
      consumes:
      - '*/*'
      description: Add secrets for source type s3
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Secret
        in: body
        name: secret
        required: true
        schema:
          $ref: '#/definitions/models.S3SecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add secrets for source type s3
      tags:
      - Secret
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Header for logged in user format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"